extract panel logic into navigation validation file
